WildlifeCompliance: refactor/remove unnecessary references to referrals
