+++ wildlifecompliance/frontend/wildlifecompliance/src/components/internal/applications/application.vue
@@ -1556,114 +1556,10 @@ export default {
         initialiseSelects: function(){
             let vm = this;
             if (!vm.initialisedSelects){
-                $(vm.$refs.department_users).select2({
-                    "theme": "bootstrap",
-                    allowClear: true,
-                    placeholder:"Select Referral"
-                }).
-                on("select2:select",function (e) {
-                    var selected = $(e.currentTarget);
-                    vm.selected_referral = selected.val();
-                }).
-                on("select2:unselect",function (e) {
-                    var selected = $(e.currentTarget);
-                    vm.selected_referral = '' 
-                });
                 vm.initialiseAssignedOfficerSelect();
                 vm.initialisedSelects = true;
             }
         },
-        // sendReferral: function(){
-        //     let vm = this;
-        //     let data = {'email':vm.selected_referral};
-        //     vm.sendingReferral = true;
-        //     vm.$http.post(helpers.add_endpoint_json(api_endpoints.applications,(vm.application.id+'/assesor_send_referral')),JSON.stringify(data),{
-        //         emulateJSON:true
-        //     }).then((response) => {
-        //         vm.sendingReferral = false;
-        //         vm.original_application = helpers.copyObject(response.body);
-        //         vm.application = response.body;
-        //         vm.application.org_applicant.address = vm.application.org_applicant.address != null ? vm.application.org_applicant.address : {};
-        //         swal(
-        //             'Referral Sent',
-        //             'The referral has been sent to '+vm.department_users.find(d => d.email == vm.selected_referral).name,
-        //             'success'
-        //         )
-        //         $(vm.$refs.department_users).val(null).trigger("change");
-        //         vm.selected_referral = '';
-        //     }, (error) => {
-        //         console.log(error);
-        //         swal(
-        //             'Referral Error',
-        //             helpers.apiVueResourceError(error),
-        //             'error'
-        //         )
-        //         vm.sendingReferral = false;
-        //     });
-        // },
-        // remindReferral:function(r){
-        //     let vm = this;
-            
-        //     vm.$http.get(helpers.add_endpoint_json(api_endpoints.referrals,r.id+'/remind')).then(response => {
-        //         vm.original_application = helpers.copyObject(response.body);
-        //         vm.application = response.body;
-        //         vm.application.org_applicant.address = vm.application.org_applicant.address != null ? vm.application.org_applicant.address : {};
-        //         swal(
-        //             'Referral Reminder',
-        //             'A reminder has been sent to '+r.referral,
-        //             'success'
-        //         )
-        //     },
-        //     error => {
-        //         swal(
-        //             'Application Error',
-        //             helpers.apiVueResourceError(error),
-        //             'error'
-        //         )
-        //     });
-        // },
-        // resendReferral:function(r){
-        //     let vm = this;
-            
-        //     vm.$http.get(helpers.add_endpoint_json(api_endpoints.referrals,r.id+'/resend')).then(response => {
-        //         vm.original_application = helpers.copyObject(response.body);
-        //         vm.application = response.body;
-        //         vm.application.org_applicant.address = vm.application.org_applicant.address != null ? vm.application.org_applicant.address : {};
-        //         swal(
-        //             'Referral Resent',
-        //             'The referral has been resent to '+r.referral,
-        //             'success'
-        //         )
-        //     },
-        //     error => {
-        //         swal(
-        //             'Application Error',
-        //             helpers.apiVueResourceError(error),
-        //             'error'
-        //         )
-        //     });
-        // },
-        // recallReferral:function(r){
-        //     let vm = this;
-            
-        //     vm.$http.get(helpers.add_endpoint_json(api_endpoints.referrals,r.id+'/recall')).then(response => {
-        //         vm.original_application = helpers.copyObject(response.body);
-        //         vm.application = response.body;
-        //         vm.application.org_applicant.address = vm.application.org_applicant.address != null ? vm.application.org_applicant.address : {};
-        //         swal(
-        //             'Referral Recall',
-        //             'The referall has been recalled from '+r.referral,
-        //             'success'
-        //         )
-        //     },
-        //     error => {
-        //         swal(
-        //             'Application Error',
-        //             helpers.apiVueResourceError(error),
-        //             'error'
-        //         )
-        //     });
-        // },
     },
     mounted: function() {
         let vm = this;
